refactor: 💡 modified field group and field group set relation
This commit will field group and field group set relationship
