+++ src/modules/form/form.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FieldGroupService } from '../maintenance/field-group/services/field-group.service';
-import { FieldGroupSetService } from './services/field-groupset.service';
+import { FieldGroupSetService } from '../maintenance/field-group-set/services/field-groupset.service';
 import { FieldInputTypeService } from '../maintenance/field-input-type/services/field-input-type.service';
 import { FieldDataTypeService } from '../maintenance/field-data-type/services/field-datatype.service';
 import { FieldOptionService } from '../maintenance/field-option/services/field-option.service';
@@ -18,7 +18,7 @@ import { FormVisibleFieldService } from './services/form-visible-fields.service'
 import { FormSectionFieldMemberService } from './services/formsection-fieldmembers.service';
 import { FieldDatatypeController } from '../maintenance/field-data-type/controllers/field-datatype.controller';
 import { FieldGroupController } from '../maintenance/field-group/controllers/field-group.controller';
-import { FieldGroupSetController } from './controllers/field-groupset.controller';
+import { FieldGroupSetController } from '../maintenance/field-group-set/controllers/field-groupset.controller';
 import { FieldInputTypeController } from '../maintenance/field-input-type/controllers/field-input-type.controller';
 import { FieldOptionChildrenController } from './controllers/field-option-children.controller';
 import { FieldOptionGroupController } from './controllers/field-option-group.controller';
@@ -33,7 +33,7 @@ import { FormController } from './controllers/form.controller';
 import { FormSectionFieldMemberController } from './controllers/formsection-fieldmembers.controller';
 import { FieldDataType } from '../maintenance/field-data-type/entities/field-datatype.entity';
 import { FieldGroup } from '../maintenance/field-group/entities/field-group.entity';
-import { FieldGroupSet } from './entities/field-groupset.entity';
+import { FieldGroupSet } from '../maintenance/field-group-set/entities/field-groupset.entity';
 import { FieldInputType } from '../maintenance/field-input-type/entities/field-input-type.entity';
 import { FieldOptionChildren } from './entities/field-option-children.entity';
 import { FieldOptionGroup } from './entities/field-option-group.entity';
