+++ src/modules/maintenance/field-group-set/controllers/field-groupset.controller.ts
@@ -4,10 +4,8 @@ import { BaseController } from 'src/core/controllers/base.contoller';
 import { FieldGroupSetService } from '../services/field-groupset.service';
 
 @Controller('api/' + FieldGroupSet.plural)
-export class FieldGroupSetController extends BaseController<
-    FieldGroupSet
-> {
-    constructor(service: FieldGroupSetService) {
-        super(service, FieldGroupSet);
+export class FieldGroupSetController extends BaseController<FieldGroupSet> {
+    constructor(fieldGroupSetService: FieldGroupSetService) {
+        super(fieldGroupSetService, FieldGroupSet);
     }
 }
