+++ src/modules/maintenance/field-group-set/entities/field-groupset.entity.ts
@@ -1,23 +1,25 @@
-
 import { Column, Entity, JoinTable, ManyToMany } from 'typeorm';
-import { FieldGroup } from '../../maintenance/field-group/entities/field-group.entity';
+import { FieldGroup } from '../../field-group/entities/field-group.entity';
 import { EntityCoreProps } from 'src/core/entities/entity-core-props';
 
 @Entity('fieldgroupset', { schema: 'public' })
 export class FieldGroupSet extends EntityCoreProps {
-
   static plural = 'fieldGroupSets';
 
-  @Column('integer', {
-    nullable: false,
-    primary: true,
-    name: 'fieldgroupsetid',
-  })
-  id: number;
-
+  /**
+   * Many To Many Relationship: FieldGroupSet and FieldGroup Entities
+   */
   @ManyToMany(type => FieldGroup, fieldGroup => fieldGroup.fieldGroupSets, {
     nullable: false,
+    cascade: true,
+    eager: true,
+    onUpdate: 'CASCADE',
+    onDelete: 'CASCADE',
   })
-  @JoinTable({ name: 'fieldgroupsetmembers' })
+  @JoinTable({
+    name: 'fieldgroupsetmembers',
+    joinColumn: { referencedColumnName: 'uid' },
+    inverseJoinColumn: { referencedColumnName: 'uid' },
+  })
   fieldGroups: FieldGroup[];
 }
