+++ src/modules/maintenance/field-group-set/field-group-set.module.ts
@@ -1,0 +1,4 @@
+import { Module } from '@nestjs/common';
+
+@Module({})
+export class FieldGroupSetModule {}
