+++ src/modules/maintenance/field-group-set/services/field-groupset.service.ts
@@ -8,8 +8,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 export class FieldGroupSetService extends BaseService<FieldGroupSet> {
     constructor(
         @InjectRepository(FieldGroupSet)
-        repository: Repository<FieldGroupSet>,
+        fieldGroupSetRepository: Repository<FieldGroupSet>,
     ) {
-        super(repository, FieldGroupSet);
+        super(fieldGroupSetRepository, FieldGroupSet);
     }
 }
